% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival.R
\name{fit_km}
\alias{fit_km}
\title{Fit Kaplan-Meier survival curves}
\usage{
fit_km(survival_data, time, event, group)
}
\arguments{
\item{survival_data}{Data frame with time, event and grouping columns.}

\item{time}{Character. Column name of the follow-up time.}

\item{event}{Character. Column name of the event indicator
(0/1 or FALSE/TRUE).}

\item{group}{Character. Column name of the grouping variable.}
}
\value{
A \code{survfit} object.
}
\description{
Fits \code{survfit(Surv(time, event) ~ group, data = ...)} and returns the
fitted object for use with \code{\link[=plot_km_curves]{plot_km_curves()}} or base plot methods.
}
\examples{
fit <- fit_km(example_survival, time = "OS", event = "status",
              group = "cluster")
fit
}
