% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_utils.R
\name{maf_common_columns}
\alias{maf_common_columns}
\title{Common column names shared by several MAF data frames}
\usage{
maf_common_columns(maf_list)
}
\arguments{
\item{maf_list}{List of MAF objects or MAF-like data frames.}
}
\value{
Character vector of common column names.
}
\description{
Returns the columns that are present in all MAF data frames of a list.
Useful before merging MAF data coming from different cohorts.
}
\examples{
common <- maf_common_columns(list(example_maf, example_maf))
head(common)
}
